Create registerFocusHotkeys.js
Move hotkeys to focus items
