+++ assets/js/hotkeys.js
@@ -1,5 +1,5 @@
 import { detectOS, isMobile } from "./detectOS.js";
-import { moveFocusToPreviousItem, moveFocusToNextItem, focusedItemIndex, resetFocusedItemIndex } from "./focusNavigation.js";
+import { setupFocusHotkeys } from "./setupFocusHotkeys.js";
 
 document.addEventListener("DOMContentLoaded", function () {
   // setup to search
@@ -33,47 +33,14 @@ document.addEventListener("DOMContentLoaded", function () {
     }
   }
 
-  // setup to move focus to next/previous item
-  const listItemLinks = [];
-  const taxonomy__indexes = document.querySelectorAll(".taxonomy__index");
-  const entriesLists = document.querySelectorAll(".entries-list");
-
-  if (taxonomy__indexes.length > 0 || entriesLists.length > 0) {
-    if (taxonomy__indexes.length > 0) {
-      taxonomy__indexes.forEach((index) => {
-        index.querySelectorAll("li > a").forEach((item) => {
-          listItemLinks.push(item);
-        });
-      });
-    }
-
-    if (entriesLists.length > 0) {
-      entriesLists.forEach((list) => {
-        list.querySelectorAll(".list__item h2 > a").forEach((item) => {
-          listItemLinks.push(item);
-        });
-      });
-    }
-
-    console.log(listItemLinks);
-
-    listItemLinks.forEach((link, index) => {
-      link.addEventListener("focusout", (_event) => {
-        if (focusedItemIndex === index) {
-          resetFocusedItemIndex();
-        }
-        console.log(`index link ${index} is unfocused`);
-        console.log(`focusedItemIndex is ${focusedItemIndex}`);
-      });
-    });
-
-    hotkeys("j", () => moveFocusToNextItem(listItemLinks));
-    hotkeys("k", () => moveFocusToPreviousItem(listItemLinks));
-  }
+  // Register focus hotkeys
+  setupFocusHotkeys();
 
   // setup to open hotkeys modal
   const hotkeysModal = document.getElementById("hotkeys-modal");
-  const hotkeysModalCloseBtn = document.getElementById("hotkeys-modal-close-btn");
+  const hotkeysModalCloseBtn = document.getElementById(
+    "hotkeys-modal-close-btn"
+  );
 
   // Close modal
   hotkeysModalCloseBtn.addEventListener("click", () => hotkeysModal.close());
